pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-ops-triage-openenv"
version = "1.0.0"
description = "OpenEnv customer support triage benchmark with deterministic graders"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Hackathon Participant" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "pydantic>=2.11.7",
  "openai>=2.7.2",
  "requests>=2.32.0",
  "PyYAML>=6.0.2",
  "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "server*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
